requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.2
joblib==1.4.2
mlflow==2.13.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pyyaml==6.0.1
httpx==0.27.0
